Move markings tab components to one folder
