+++ src/panels/NightSkyPanel/tabs/MarkingsTab/NightSkyMarkingBox.tsx
