+++ src/panels/NightSkyPanel/tabs/MarkingsTab/ToggleCard.tsx
